cleaned up day 03; readme filled with results

+++ src/year2025/day02.rs
@@ -1,6 +1,6 @@
 //! # Gift Shop
 //!
-//!
+//! Part 2 must be refactored the same as Part 1 to get rid of string usage.
 
 pub fn parse(input: &str) -> &str {
     input.trim()
@@ -16,7 +16,7 @@ pub fn part1(input: &str) -> u64 {
 
         for num in start..=end {
             let n_digits = num.ilog10() + 1;
-            let mask: u64 = 10u64.pow((n_digits / 2) as u32);
+            let mask: u64 = 10u64.pow(n_digits / 2);
 
             let left = num / mask;
             let right = num % mask;
